feat: v1-v2 positions
