+++ packages/api/src/app.ts
@@ -8,9 +8,9 @@ import {
   getTimestampInSeconds,
   getTrustedProxies,
 } from '@voltz-protocol/commons-v2';
-import { getAmm } from './v1-queries/common/getAMM';
-import { getPortfolioPositionDetails } from './v1-queries/get-position-details/getPortfolioPositionDetails';
-import { getPositionPnL } from './v1-queries/position-pnl/getPositionPnL';
+import { getAmm } from './old-v1-queries/common/getAMM';
+import { getPortfolioPositionDetails } from './old-v1-queries/get-position-details/getPortfolioPositionDetails';
+import { getPositionPnL } from './old-v1-queries/position-pnl/getPositionPnL';
 import {
   pullAllChainPools,
   getFixedRates,
@@ -22,7 +22,7 @@ import {
   getVoyages,
   getWalletVoyages,
 } from '@voltz-protocol/indexer-v1';
-import { getPortfolioPositions as getPortfolioPositionsV1 } from './v1-queries/portfolio-positions/getPortfolioPositions';
+import { getPortfolioPositions as getPortfolioPositionsV1 } from './old-v1-queries/portfolio-positions/getPortfolioPositions';
 import { getApyFromTo, getLiquidityIndexAt } from '@voltz-protocol/bigquery-v2';
 import { getV1V2Pools } from './v1v2-queries/get-pools/getV1V2Pools';
 import { getV1V2PortfolioPositions } from './v1v2-queries/get-portfolio-positions/getPortfolioPositions';
