+++ packages/api/src/old-v1-queries/get-position-details/synthetisizeHistory.ts
