+++ packages/api/src/v1-queries/get-pools/getV1Pools.ts
@@ -1,13 +1,13 @@
 import { SupportedChainId } from '@voltz-protocol/commons-v2';
 import { V1Pool } from './types';
 import { pullAllChainPools } from '@voltz-protocol/indexer-v1';
-import { buildPool } from './buildPool';
+import { buildV1Pool } from './buildV1Pool';
 
 export const getV1Pools = async (
   chainIds: SupportedChainId[],
 ): Promise<V1Pool[]> => {
   const rawPools = await pullAllChainPools(chainIds);
-  const promises = rawPools.map(buildPool);
+  const promises = rawPools.map(buildV1Pool);
   const responses = await Promise.allSettled(promises);
 
   const pools = responses.map((r) => {
