+++ packages/api/src/v1-queries/get-position-details/getPortfolioPositionDetails.ts
@@ -1,15 +1,8 @@
 import { getPositions as getRawPositions } from '@voltz-protocol/subgraph-data';
 
-import { PortfolioPositionAMM } from '../portfolio-positions/types';
 import { synthetisizeHistory } from './synthetisizeHistory';
 import { PortfolioPositionDetails } from './types';
-import {
-  tickToFixedRate,
-  getTokenPriceInUSD,
-  SECONDS_IN_YEAR,
-  getProtocolName,
-  isBorrowingProtocol,
-} from '@voltz-protocol/commons-v2';
+import { tickToFixedRate, SECONDS_IN_YEAR } from '@voltz-protocol/commons-v2';
 import {
   getPositionInfo,
   getVariableFactor,
@@ -17,6 +10,7 @@ import {
 } from '@voltz-protocol/indexer-v1';
 import { getPositionPnL } from '../position-pnl/getPositionPnL';
 import { getSubgraphURL } from '../subgraph/getSubgraphURL';
+import { getV1Pool } from '../get-pools/getV1Pool';
 
 const decodePositionId = (
   positionId: string,
@@ -41,6 +35,7 @@ type GetPortfolioPositionDetails = {
   positionId: string;
   includeHistory: boolean;
 };
+
 export const getPortfolioPositionDetails = async ({
   positionId,
   includeHistory,
@@ -50,11 +45,9 @@ export const getPortfolioPositionDetails = async ({
   const { chainId, vammAddress, ownerAddress, tickLower, tickUpper } =
     decodePositionId(positionId);
 
-  const fixLow = tickToFixedRate(tickUpper);
-  const fixHigh = tickToFixedRate(tickLower);
-
   const subgraphURL = getSubgraphURL(chainId);
-  // Get transaction history
+
+  // Fetch positions from subgraph
   const positions = subgraphURL
     ? (
         await getRawPositions(
@@ -69,11 +62,16 @@ export const getPortfolioPositionDetails = async ({
       ).filter((p) => p.tickLower === tickLower && p.tickUpper === tickUpper)
     : [];
 
+  // Check the number of positions to be 1
   if (positions.length === 0 || positions.length >= 2) {
     throw new Error('No position');
   }
 
+  // Process the found position
   const position = positions[0];
+
+  const fixLow = tickToFixedRate(tickUpper);
+  const fixHigh = tickToFixedRate(tickLower);
 
   const positionType =
     position.positionType === 3
@@ -81,33 +79,16 @@ export const getPortfolioPositionDetails = async ({
       : position.positionType === 2
       ? 'Variable'
       : 'Fixed';
-  const tokenName = position.amm.tokenName;
 
   const txs = synthetisizeHistory(position);
 
-  const amm: PortfolioPositionAMM = {
-    id: vammAddress,
-    chainId,
-
-    marginEngineAddress: position.amm.marginEngineId,
-    isV2: false,
-    isBorrowing: isBorrowingProtocol(position.amm.protocolId),
-    market: getProtocolName(position.amm.protocolId),
-
-    rateOracle: {
-      address: position.amm.rateOracleId,
-      protocolId: position.amm.protocolId,
-    },
-
-    underlyingToken: {
-      address: position.amm.tokenId,
-      name: tokenName.toLowerCase() as 'eth' | 'usdc' | 'usdt' | 'dai',
-      tokenDecimals: position.amm.tokenDecimals,
-    },
-
-    termStartTimestampInMS: position.amm.termStartTimestampInMS,
-    termEndTimestampInMS: position.amm.termEndTimestampInMS,
-  };
+  const amm = await getV1Pool(chainId, vammAddress);
+  if (!amm) {
+    throw new Error(
+      `Could not find pool (in BigQuery) for ${chainId}-${vammAddress}`,
+    );
+  }
+  const tokenPriceUSD = amm.underlyingToken.priceUSD;
 
   // Get fresh information about the position
   const {
@@ -127,8 +108,6 @@ export const getPortfolioPositionDetails = async ({
   );
 
   const notional = positionType === 'LP' ? notionalProvided : notionalTraded;
-
-  const tokenPriceUSD = await getTokenPriceInUSD(position.amm.tokenName);
 
   if (position.isSettled) {
     if (position.settlements.length === 0 || position.settlements.length >= 2) {
