+++ packages/api/src/v1v2-queries/get-pools/types.ts
@@ -1,4 +1,4 @@
-import { V1Pool } from '../../v1-queries-new/get-pools/types';
+import { V1Pool } from '../../v1-queries/get-pools/types';
 import { V2Pool } from '../../v2-queries/get-pools/types';
 
 export type V1V2Pool = V1Pool & V2Pool;
