+++ packages/api/src/v1v2-queries/get-portfolio-positions/extendV2Position.ts
@@ -1,10 +1,10 @@
-import { PortfolioPositionV2 } from '../../v2-queries/get-portfolio-positions/types';
+import { V2PortfolioPosition } from '../../v2-queries/get-portfolio-positions/types';
 import { extendV2Pool } from '../get-pools/extendV2Pool';
-import { PortfolioPositionV1V2 } from './types';
+import { V1V2PortfolioPosition } from './types';
 
 export const extendV2Position = (
-  p: PortfolioPositionV2,
-): PortfolioPositionV1V2 => ({
+  p: V2PortfolioPosition,
+): V1V2PortfolioPosition => ({
   ...p,
   amm: extendV2Pool(p.amm),
 });
