+++ packages/api/src/v1v2-queries/get-portfolio-positions/types.ts
@@ -1,9 +1,9 @@
-import { V1Pool } from '../../v1-queries-new/get-pools/types';
-import { PortfolioPositionV1 } from '../../v1-queries-new/get-portfolio-positions/types';
+import { V1Pool } from '../../v1-queries/get-pools/types';
+import { V1PortfolioPosition } from '../../v1-queries/get-portfolio-positions/types';
 import { V2Pool } from '../../v2-queries/get-pools/types';
-import { PortfolioPositionV2 } from '../../v2-queries/get-portfolio-positions/types';
+import { V2PortfolioPosition } from '../../v2-queries/get-portfolio-positions/types';
 
-export type PortfolioPositionV1V2 = PortfolioPositionV1 &
-  PortfolioPositionV2 & {
+export type V1V2PortfolioPosition = V1PortfolioPosition &
+  V2PortfolioPosition & {
     amm: V1Pool & V2Pool;
   };
